Created an initial `File` implementation and read from disk
